Add support for nested requirements
This patch implements the support for nested requirements.

The use case for this is when a project contains a requirements.txt
file pointing to another requirements file (i.e.:
production-requirements.txt) which also points to other requirement
files.

The patch achieves this by compiling the list of all the requirements,
hashing it, and comparing this hash to the hash of the previous tox run
(if any).