Fix file names of files with previous hashes (#7)
`tox-battery` stores a hash per requirements file and per environment to
check for changes. There was a subtle bug, which would result in the
environment not to be included in the file name, resulting in the same
file name for multiple environments, which broke the functionality of
`tox-battery` when using it with multiple environments which used
requirement files with the same name. This commit fixes that.