Support pip 10
Problem
-------
As expected when using the internal implementation on major changes it breaks.
With release of pip 10 functionality to parse the requirements files is moved to
a different place so it can't be imported any more.

Solution
--------
Fallback to a new location if an old one cannot be found.

Testing
-------
All the tests pass